Added admin discount and creator to bookings
