+++ .pre-commit-config.yaml
@@ -61,4 +61,4 @@ repos:
     rev: v0.812
     hooks:
       - id: mypy
-        additional_dependencies: [django-stubs,dj-database-url,django-filter,django-cors-headers,django-autoslug,graphene-django,psycopg2-binary,django-graphql-auth,gunicorn,shortuuid,squareup,django-environ]
+        additional_dependencies: [django-stubs,dj-database-url,django-filter,django-cors-headers,django-autoslug,graphene-django,psycopg2-binary,django-graphql-auth,gunicorn,shortuuid,squareup,django-environ,django-guardian]
