+++ uobtheatre/utils/test_utils.py
@@ -2,7 +2,84 @@
 Utils for the tests
 """
 
+import pytest
+
 from uobtheatre.bookings.models import Ticket
+from uobtheatre.utils.utils import combinations
+
+
+@pytest.mark.parametrize(
+    "inputs, length, output",
+    [
+        (
+            [1, 2, 3],
+            2,
+            [
+                (1,),
+                (2,),
+                (3,),
+                (1, 1),
+                (1, 2),
+                (1, 3),
+                (2, 1),
+                (2, 2),
+                (2, 3),
+                (3, 1),
+                (3, 2),
+                (3, 3),
+            ],
+        ),
+        (
+            [1, 2, 3],
+            3,
+            [
+                (1,),
+                (2,),
+                (3,),
+                (1, 1),
+                (1, 2),
+                (1, 3),
+                (2, 1),
+                (2, 2),
+                (2, 3),
+                (3, 1),
+                (3, 2),
+                (3, 3),
+                (1, 1, 1),
+                (1, 1, 2),
+                (1, 1, 3),
+                (1, 2, 1),
+                (1, 2, 2),
+                (1, 2, 3),
+                (1, 3, 1),
+                (1, 3, 2),
+                (1, 3, 3),
+                (2, 1, 1),
+                (2, 1, 2),
+                (2, 1, 3),
+                (2, 2, 1),
+                (2, 2, 2),
+                (2, 2, 3),
+                (2, 3, 1),
+                (2, 3, 2),
+                (2, 3, 3),
+                (3, 1, 1),
+                (3, 1, 2),
+                (3, 1, 3),
+                (3, 2, 1),
+                (3, 2, 2),
+                (3, 2, 3),
+                (3, 3, 1),
+                (3, 3, 2),
+                (3, 3, 3),
+            ],
+        ),
+    ],
+)
+def test_combinations(inputs, length, output):
+    calculated_combinations = combinations(inputs, length)
+    assert set(calculated_combinations) == set(output)
+    assert len(calculated_combinations) == len(output)
 
 
 def ticket_list_dict_gen(ticket_list):
